requirements.txt
pydantic
openai-agents
bs4
requests
tavily-python
python-dotenv
rich
